Consolidate the Offset classes
