+++ akka-docs/rst/java/code/docs/persistence/query/LeveldbPersistenceQueryDocTest.java
@@ -12,6 +12,7 @@ import akka.actor.ActorSystem;
 import akka.persistence.journal.WriteEventAdapter;
 import akka.persistence.journal.Tagged;
 import akka.persistence.query.EventEnvelope;
+import akka.persistence.query.Sequence;
 import akka.persistence.query.javadsl.*;
 import akka.persistence.query.PersistenceQuery;
 import akka.persistence.query.journal.leveldb.javadsl.LeveldbReadJournal;
