+++ akka-persistence-query/src/main/scala/akka/persistence/query/Offset.scala
@@ -2,16 +2,15 @@
  * Copyright (C) 2015-2016 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.persistence.query.scaladsl
+package akka.persistence.query
 
 import java.util.UUID
-import akka.persistence.query.javadsl
 
-trait Offset extends javadsl.Offset {
+trait Offset {
   def hashCode: Int
 }
 
-final case class Sequence(override val value: Long) extends javadsl.Sequence(value) with Offset with Ordered[Sequence] {
+final case class Sequence(val value: Long) extends Offset with Ordered[Sequence] {
 
   override def hashCode: Int = value.hashCode
 
@@ -20,7 +19,7 @@ final case class Sequence(override val value: Long) extends javadsl.Sequence(val
   override def compare(that: Sequence): Int = value.compare(that.value)
 }
 
-final case class TimeBasedUUID(override val value: UUID) extends javadsl.TimeBasedUUID(value) with Offset with Ordered[TimeBasedUUID] {
+final case class TimeBasedUUID(val value: UUID) extends Offset with Ordered[TimeBasedUUID] {
   if (value == null || value.version != 1) {
     throw new IllegalArgumentException("UUID " + value + " is not a time-based UUID")
   }
@@ -36,6 +35,4 @@ final case object NoOffset extends Offset {
   override def hashCode: Int = 0
 
   override def toString: String = "NoOffset"
-
-  override def asScala = NoOffset
 }
